色々

+++ math/binary_search_double.hpp
@@ -5,11 +5,11 @@
  */
 
 template<typename F>
-double bs(double mn,double mx,F func) {
-    double left = mn;
-    double right =mx;
+long double bs(long double mn,long double mx,F func) {
+    long double left = mn;
+    long double right =mx;
     for(int i=0;i<100;i++){
-        double mid=(right+left)*0.5;
+        long double mid=(right+left)*0.5;
         if (!func(mid)) right = mid;
         else left = mid;
     }
